+++ raiden-dapp/CHANGELOG.md
@@ -10,7 +10,10 @@
 - [#693] Localized error messages for Pathfinding errors.
 
 ### Fixed
+- [#1067] Fixed displayed service fee on summary screen.
 
+
+[#1067]: https://github.com/raiden-network/light-client/issues/1067
 [#1071]: https://github.com/raiden-network/light-client/issues/1071
 [#1015]: https://github.com/raiden-network/light-client/issues/1015
 [#695]: https://github.com/raiden-network/light-client/issues/695